Enviando versão 5.0
Com validações no servidor.
Casos de uso Custodiais .
Login do ADM e jornaleiro

+++ Sistema-de-Notas-Digitais/api/modulos/NotaVenda/ControladoraNotaVenda.php
@@ -27,21 +27,27 @@
 
 			function criarNotaVenda( $params ){
 				try{
-					$pontoVenda = $this->pontoVendaDAO->comId( $params['pontoVenda'] );
+
+					if( is_numeric($params['pontoVenda'])){
+						$pontoVenda = $this->pontoVendaDAO->comId( $params['pontoVenda'] );
+					}
+					else{ 
+						throw new DAOException("Forneca o ponto de Venda corretamente", 1);
+					}
+
 					$dataPgmt = $this->calcularPagamento( $pontoVenda->getJornaleiro() );
 					$data =  date('d/m/Y');
-				/*	$data = explode("/", $data );
-					$data = implode("", $data );
-					var_dump($data);*/
 					$itensNotaForm = $params['itensNota'];
-
 					$itensNota = [];
 					$numeroDeJornais =(int) $this->jornalDAO->numeroDeRegistros();
 					for( $contador = 0 ; $contador <  $numeroDeJornais ; $contador = $contador + 1 ) {
-
-						$precoCapa = $this->precoCapaDAO->comId( $itensNotaForm['precosCapa'][ $contador ] );
-						$itemNota = new ItemNota( $itensNotaForm['qtdEntregue'][$contador], 0 , $precoCapa );
-						$itensNota[] = $itemNota ;
+							if(is_numeric($itensNotaForm['precosCapa'][ $contador ]) && is_numeric($itensNotaForm['qtdEntregue'][$contador]) ){
+								$precoCapa = $this->precoCapaDAO->comId( $itensNotaForm['precosCapa'][ $contador ] );
+								$itemNota = new ItemNota( $itensNotaForm['qtdEntregue'][$contador], 0 , $precoCapa );
+								$itensNota[] = $itemNota ;
+							}
+							else{ throw new DAOException("Os precos e as quantidades devem ser numericos", 1);
+							}
 					}
 				
 					$notaVenda = new NotaVenda( $data, $dataPgmt, 0.0, $pontoVenda );
@@ -50,7 +56,7 @@
 
 					return $this->geradoraResposta->criado( '',GeradoraResposta::TIPO_TEXTO);	
 				}catch( DAOException $e ){
-
+					return $this->geradoraResposta->erro( $e->getMessage(), GeradoraResposta::TIPO_TEXTO);
 				}	
 			}
 
@@ -59,14 +65,19 @@
 			function registrarPagamento( $params ){
 				try{
 					$id_pontovenda = $params['pontoVenda'];
-					$pontoVenda = $this->pontoVendaDAO->comId( $id_pontovenda );
-					$dataNota = $params['dataNota'];
-					$notaVenda = $this->notaVendaDAO->buscarPorPontoEData( $pontoVenda, $dataNota );
-					$notaVenda->setPaga(1);
-					$this->notaVendaDAO->registrarPagamento( $notaVenda );
-					return $this->geradoraResposta->semConteudo();
+					if( is_numeric( $id_pontovenda ) ){
+						$pontoVenda = $this->pontoVendaDAO->comId( $id_pontovenda );
+						$dataNota = $params['dataNota'];
+						$notaVenda = $this->notaVendaDAO->buscarPorPontoEData( $pontoVenda, $dataNota );
+						$notaVenda->setPaga(1);
+						$this->notaVendaDAO->registrarPagamento( $notaVenda );
+						return $this->geradoraResposta->semConteudo();
+					}
+					else{
+						throw new DAOException("Forneca o ponto de Venda Corretamente");
+					}
 				}catch( DAOException $e ){
-
+					return $this->geradoraResposta->erro( $e->getMessage(), GeradoraResposta::TIPO_TEXTO);
 				}
 
 			}
@@ -74,31 +85,52 @@
 			function registrarVenda( $params ){
 				try{
 					$id_notaVenda = $params['id'];
-					$notaVenda = $this->notaVendaDAO->comId( $id_notaVenda );
-					$itensNota = $notaVenda->getItensNota() ;
-					$numeroDeJornais =(int) $this->jornalDAO->numeroDeRegistros();
-					$itensForm =  $params["itensNota"] ;
-					for( $contador = 0 ; $contador < $numeroDeJornais ; $contador ++){
-						$row =  $itensForm[$contador] ;
-						$itemNota = $itensNota[$contador] ;
-						$itemNota->setQtdVendido( $row['qtdVendido']);
-						$itensNota[$contador] = $itemNota ;
+					if( is_numeric($id_notaVenda)){
+						$notaVenda = $this->notaVendaDAO->comId( $id_notaVenda );
+						$itensNota = $notaVenda->getItensNota() ;
+						$numeroDeJornais =(int) $this->jornalDAO->numeroDeRegistros();
+						$itensForm =  $params["itensNota"] ;
+						for( $contador = 0 ; $contador < $numeroDeJornais ; $contador ++){
+							$row =  $itensForm[$contador] ;
+							if( is_numeric($row['qtdVendido']) ){
+								$itemNota = $itensNota[$contador] ;
+								if( $itemNota->getQtdEntregue() >= $row['qtdVendido']){
+									$itemNota->setQtdVendido( $row['qtdVendido']);
+									$itensNota[$contador] = $itemNota ;
+								}
+								else{
+									throw new Exception("A quantidade de vendidos nao pode ser maior que a quantidade de entregues");
+								}
+							}
+							else{
+								throw new DAOException("A quantidade vendida deve ser um numero !");
+							}
+						}
+						$notaVenda->setItensNota( $itensNota );
+						$this->notaVendaDAO->registrarVenda( $notaVenda );
+						return $this->geradoraResposta->semConteudo();
 					}
-					$notaVenda->setItensNota( $itensNota );
-					$this->notaVendaDAO->registrarVenda( $notaVenda );
-					return $this->geradoraResposta->semConteudo();
+					else{
+						throw new DAOException("Nota de Venda incorreta");
+						
+					}
 				}catch( DAOException $e ){
-
+					return $this->geradoraResposta->erro( $e->getMessage(), GeradoraResposta::TIPO_TEXTO);
 				}
 
 			}
 
 			function buscarComId( $id ){
 				try{
-					$notaVenda = $this->notaVendaDAO->comId( $id );
-					return $this->geradoraResposta->ok( $notaVenda,GeradoraResposta::TIPO_JSON  );
+					if( is_numeric($id )){
+						$notaVenda = $this->notaVendaDAO->comId( $id );
+						return $this->geradoraResposta->ok( $notaVenda,GeradoraResposta::TIPO_JSON);
+					}
+					else{
+						throw new Exception("Erro ao buscar a Nota de Venda");
+					}
 				}catch( DAOException $e ){
-
+					return $this->geradoraResposta->erro( $e->getMessage(), GeradoraResposta::TIPO_TEXTO);
 				}
 			}
 
@@ -106,37 +138,61 @@
 				try{
 
 					$id_pontovenda = $params['pontoVenda'];
-					$pontoVenda = $this->pontoVendaDAO->comId( (int)$id_pontovenda);
+					if( is_numeric( $id_pontovenda )){
+						$pontoVenda = $this->pontoVendaDAO->comId( (int)$id_pontovenda);
+					}
+					else{ throw new Exception("Erro ao buscar a Nota de Venda");}
 					$dataNota = $params['dataNota'];
-					$notaVenda = $this->notaVendaDAO->buscarPorPontoEData( $pontoVenda, $dataNota ); 
-
-					return $this->geradoraResposta->ok( $notaVenda, GeradoraResposta::TIPO_JSON );
-				}catch(DAOException $e ){
+					if( preg_match(self::ER_DATA , $dataNota  )){	
+						$notaVenda = $this->notaVendaDAO->buscarPorPontoEData( $pontoVenda, $dataNota ); 
+						return $this->geradoraResposta->ok( $notaVenda, GeradoraResposta::TIPO_JSON );
+					}
+					else{ throw new Exception("Erro ao buscar a Nota de Venda"); }
 					
+				}catch(DAOException $e ){
+					return $this->geradoraResposta->erro( $e->getMessage(), GeradoraResposta::TIPO_TEXTO);
 				}
 			}
 
 			function buscarPorData( $params ){
 				try{
-
 					$dataNota = $params['dataNota'];
+				if( preg_match( self::ER_DATA , $dataNota  )){
 					$notasVenda = $this->notaVendaDAO->buscarPorData( $dataNota ); 
-
 					return $this->geradoraResposta->ok( $notasVenda, GeradoraResposta::TIPO_JSON );
-				}catch(DAOException $e ){
+				}
+				else{
+						throw new DAOException("Forneca a data Corretamente");
 					
+					}			
+				}catch(DAOException $e ){
+					return $this->geradoraResposta->erro( $e->getMessage(), GeradoraResposta::TIPO_TEXTO);
 				}
 			}
 
 			function buscarPorDataENaoPaga( $params ){
 				try{
-
 					$dataNota = $params['dataNota'];
-					$notasVenda = $this->notaVendaDAO->buscarPorDataENaoPaga( $dataNota ); 
-
-					return $this->geradoraResposta->ok( $notasVenda, GeradoraResposta::TIPO_JSON );
+					if( preg_match( self::ER_DATA , $dataNota  )){
+						
+						$notasVenda = $this->notaVendaDAO->buscarPorDataENaoPaga( $dataNota ); 
+						return $this->geradoraResposta->ok( $notasVenda, GeradoraResposta::TIPO_JSON );
+					}
+					else{
+						throw new DAOException("Forneca a data Corretamente");
+					}
 				}catch(DAOException $e ){
-					
+					return $this->geradoraResposta->erro( $e->getMessage(), GeradoraResposta::TIPO_TEXTO);
+				}
+			}
+
+			function buscarNotasDoDia(){
+				try{
+					$data = [ "dataNota" => date('d/m/Y') ];
+					$this->buscarPorData( $data );
+
+				}catch( DAOException $e ){
+					return $this->geradoraResposta->erro( $e->getMessage(), GeradoraResposta::TIPO_TEXTO);
 				}
 			}
 
@@ -144,31 +200,44 @@
 				try{
 
 					$dataPgmt = $params['dataPgmt'];
-					$id = $params['id'];
-					$itensNota = [] ;
-					foreach ( $params['itensNota'] as $itemNota) {
 
-						$itensNota[] = new ItemNota( $itemNota[0], $itemNota[1], null ,(int) $itemNota[2]);
+					if( preg_match( self::ER_DATA , $dataPgmt  )){
+						$id = $params['id'];
+						if( is_numeric($id)){
+							$itensNota = [] ;
+							foreach ( $params['itensNota'] as $itemNota) {
+
+								$itensNota[] = new ItemNota( $itemNota[0], $itemNota[1], null ,(int) $itemNota[2]);
+							}
+
+							$notaVenda = new NotaVenda( null, $dataPgmt, 0, null, $id, $itensNota );
+							$this->notaVendaDAO->alterar( $notaVenda );
+						}
+						else{
+							throw new DAOException("Erro ao alterar a nota de venda");
+						}
+
 					}
 
-					$notaVenda = new NotaVenda( null, $dataPgmt, 0, null, $id, $itensNota );
-					$this->notaVendaDAO->alterar( $notaVenda );
-
+					else{
+						throw new DAOException("Forneca a data Corretamente");
+					}
 				}catch( DAOException $e ){
-
+					return $this->geradoraResposta->erro( $e->getMessage(), GeradoraResposta::TIPO_TEXTO);
 				}
 
 			}
 
 			function excluirNotaVenda( $params ){
 				try{
-
 					$id = $params['id'];
-					$notaVenda = $this->notaVendaDAO->comId( $id );
-					$this->notaVendaDAO->remover( $notaVenda );
-					
+					if( is_numeric($id )){
+						$notaVenda = $this->notaVendaDAO->comId( $id );
+						$this->notaVendaDAO->remover( $notaVenda );
+					}
+						
 				}catch( DAOException $e  ){
-
+					return $this->geradoraResposta->erro( $e->getMessage(), GeradoraResposta::TIPO_TEXTO);
 				}
 			}
 
@@ -194,15 +263,14 @@
 					return $this->geradoraResposta->ok( $pontosSemNota, GeradoraResposta::TIPO_JSON );
 
 				}catch( DAOException $e ){
-
+					return $this->geradoraResposta->erro( $e->getMessage(), GeradoraResposta::TIPO_TEXTO);
 				}
 			}
 
 			private function calcularPagamento( $jornaleiro ){
 				
 				$tipoPagamento = $jornaleiro->getTipoPagamento();
-				$dataPgmt =  date('dmy');
-				$dataPgmt = new DateTime( $dataPgmt );
+				$dataPgmt = new DateTime(null );
 				
 				switch ($tipoPagamento) {
 					case 1 :
@@ -212,9 +280,7 @@
 					case 2 : 
 							$dataPgmt->modify('+1 week');
 							break;
-					case 3 :
-							$dataPgmt->modify('+1 month');	
-							break;
+
 					default:
 							
 						break;
@@ -222,9 +288,7 @@
 				return $dataPgmt->format('d/m/Y'); ;
 			}
 
-
-
-			
+	const ER_DATA = "/^\d{2}\/\d{2}\/\d{4}$/";
 
 }
